+++ game_patch/misc/game.cpp
@@ -101,7 +101,7 @@ CodeInjection jpeg_write_bitmap_overflow_fix2{
 
 int bm_load_if_exists(const char* name, int unk, bool generate_mipmaps)
 {
-    if (rf::file_exists(name)) {
+    if (rf::File{}.find(name)) {
         return rf::bm::load(name, unk, generate_mipmaps);
     }
     return -1;