+++ docs/CHANGELOG.md
@@ -26,6 +26,7 @@ Version 1.8.1 (not released yet)
 - Hide player riot shield third person model in cutscenes
 - Fix player third person weapon model animations after loading the game from a save file
 - Change default volume to 0.5
+- Log a warning when sound file cannot be find
 
 Version 1.8.0 (released 2022-09-17)
 -----------------------------------